+++ CHANGELOG.md
@@ -3,7 +3,11 @@ A more detailed list of changes is available in the corresponding milestones for
 
 
 ## 0.7.4 (2019-May-06)
+### Dependencies
   - Removed defusedxml dependency. We were only using it for its `defused.lxml` module which is now deprecated (issue #2477)
+
+### New checks
+  - **[com.google.fonts/check/metadata/designer_values]:** We must use commas instead of forward slashes because the fonts.google.com directory will segment string to list on comma and display the first item in the list as the "principal designer" and the other items as contributors.
 
 
 ## 0.7.3 (2019-Apr-19)